requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
yfinance>=0.2.0
python-dotenv>=1.0.0
